long id support

+++ carbonj.service/src/main/java/com/demandware/carbonj/service/db/index/IndexStore.java
@@ -29,5 +29,5 @@ public interface IndexStore<K, R extends Record<K>>
 
     K maxKey();
 
-    int scan( K startKey, K endKey, Consumer<R> c );
+    long scan( K startKey, K endKey, Consumer<R> c );
 }
